refactor rxjs utils
