+++ projects/core/src/util/index.ts
@@ -5,8 +5,7 @@ export * from './glob.service';
 export * from './http-params-uri.encoder';
 export * from './normalize-http-error';
 export * from './regex-pattern';
-export * from './withdraw-on';
-export * from './get-last-value-sync';
+export * from './rxjs/index';
 export * from './time-utils';
 export * from './external-js-file-loader/index';
 export * from './ssr.tokens';
