+++ projects/core/src/util/rxjs/get-last-value-sync.ts
